refactor(ops): split basic.py into submodules aligned with C++ structure
Split ops/basic.py (2275 lines) into 9 focused modules:
- _common.py: validation helpers
- elementwise.py: add, sub, mul, div, inplace ops
- unary.py: exp, log, relu
- reduction.py: sum, mean, max, softmax
- matmul.py: matmul, transpose, linear_bias_gelu
- nn.py: gelu, silu, layernorm, rmsnorm, sdpa_*, rope_*
- embedding.py: embedding_lookup*, kv_cache_*
- sampling.py: sample_*, set_sampling_seed
- tensor.py: concat, repeat, transpose_3d, reshape, cast

basic.py now re-exports all functions for backwards compatibility.
Module structure aligns with native/ops/ directory.
