+++ src/pygpukit/ops/reduction.py
@@ -1,0 +1,176 @@
+"""Reduction operations for GPUArrays.
+
+Corresponds to native/ops/reduction/.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+
+from pygpukit.core.array import GPUArray
+from pygpukit.core.backend import NativeBackend, get_backend
+from pygpukit.core.factory import from_numpy
+from pygpukit.ops._common import _validate_float_dtype
+
+
+def sum(a: GPUArray) -> GPUArray:
+    """Sum of all elements.
+
+    Args:
+        a: Input array (float32 or float64).
+
+    Returns:
+        A scalar GPUArray (shape [1]) containing the sum.
+
+    Raises:
+        ValueError: If dtype is not float32 or float64.
+    """
+    _validate_float_dtype(a, "sum")
+
+    backend = get_backend()
+
+    if isinstance(backend, NativeBackend) and backend.is_available():
+        return _sum_native(a)
+    else:
+        return _sum_cpu(a)
+
+
+def _sum_cpu(a: GPUArray) -> GPUArray:
+    """CPU implementation of sum."""
+    a_np = a.to_numpy()
+    result_np = np.array([np.sum(a_np)], dtype=a_np.dtype)
+    return from_numpy(result_np)
+
+
+def _sum_native(a: GPUArray) -> GPUArray:
+    """Native C++ CUDA implementation of sum (zero-copy)."""
+    from pygpukit.core.backend import get_native_module
+
+    native = get_native_module()
+    a_native = a._get_native()
+    c_native = native.sum(a_native)
+    return GPUArray._wrap_native(c_native)
+
+
+def mean(a: GPUArray) -> GPUArray:
+    """Mean of all elements.
+
+    Args:
+        a: Input array (float32 or float64).
+
+    Returns:
+        A scalar GPUArray (shape [1]) containing the mean.
+
+    Raises:
+        ValueError: If dtype is not float32 or float64.
+    """
+    _validate_float_dtype(a, "mean")
+
+    backend = get_backend()
+
+    if isinstance(backend, NativeBackend) and backend.is_available():
+        return _mean_native(a)
+    else:
+        return _mean_cpu(a)
+
+
+def _mean_cpu(a: GPUArray) -> GPUArray:
+    """CPU implementation of mean."""
+    a_np = a.to_numpy()
+    result_np = np.array([np.mean(a_np)], dtype=a_np.dtype)
+    return from_numpy(result_np)
+
+
+def _mean_native(a: GPUArray) -> GPUArray:
+    """Native C++ CUDA implementation of mean (zero-copy)."""
+    from pygpukit.core.backend import get_native_module
+
+    native = get_native_module()
+    a_native = a._get_native()
+    c_native = native.mean(a_native)
+    return GPUArray._wrap_native(c_native)
+
+
+def max(a: GPUArray) -> GPUArray:
+    """Max of all elements.
+
+    Args:
+        a: Input array (float32 or float64).
+
+    Returns:
+        A scalar GPUArray (shape [1]) containing the maximum value.
+
+    Raises:
+        ValueError: If dtype is not float32 or float64.
+    """
+    _validate_float_dtype(a, "max")
+
+    backend = get_backend()
+
+    if isinstance(backend, NativeBackend) and backend.is_available():
+        return _max_native(a)
+    else:
+        return _max_cpu(a)
+
+
+def _max_cpu(a: GPUArray) -> GPUArray:
+    """CPU implementation of max."""
+    a_np = a.to_numpy()
+    result_np = np.array([np.max(a_np)], dtype=a_np.dtype)
+    return from_numpy(result_np)
+
+
+def _max_native(a: GPUArray) -> GPUArray:
+    """Native C++ CUDA implementation of max (zero-copy)."""
+    from pygpukit.core.backend import get_native_module
+
+    native = get_native_module()
+    a_native = a._get_native()
+    c_native = native.max(a_native)
+    return GPUArray._wrap_native(c_native)
+
+
+def softmax(input: GPUArray) -> GPUArray:
+    """Softmax activation applied row-wise.
+
+    Computes: y[i] = exp(x[i] - max(x)) / sum(exp(x - max(x)))
+
+    Args:
+        input: Input array of shape [batch, features].
+
+    Returns:
+        A new GPUArray containing the softmax output.
+
+    Raises:
+        ValueError: If input is not 2D or dtype is not a float type.
+    """
+    _validate_float_dtype(input, "softmax")
+
+    if input.ndim != 2:
+        raise ValueError(f"softmax expects 2D input [batch, features], got {input.ndim}D")
+
+    backend = get_backend()
+
+    if isinstance(backend, NativeBackend) and backend.is_available():
+        return _softmax_native(input)
+    else:
+        return _softmax_cpu(input)
+
+
+def _softmax_cpu(input: GPUArray) -> GPUArray:
+    """CPU implementation of softmax."""
+    x = input.to_numpy()
+    # Numerical stability: subtract max
+    x_max = x.max(axis=1, keepdims=True)
+    exp_x = np.exp(x - x_max)
+    return from_numpy(exp_x / exp_x.sum(axis=1, keepdims=True))
+
+
+def _softmax_native(input: GPUArray) -> GPUArray:
+    """Native C++ CUDA implementation of softmax (zero-copy)."""
+    from pygpukit.core.backend import get_native_module
+
+    native = get_native_module()
+    input_native = input._get_native()
+    c_native = native.softmax(input_native)
+    return GPUArray._wrap_native(c_native)
